+++ src/main/java/minechem/client/gui/GuiToggleSwitch.java
@@ -1,16 +1,18 @@
 package minechem.client.gui;
 
-import cpw.mods.fml.client.FMLClientHandler;
+import java.util.HashMap;
+
 import minechem.common.utils.MinechemHelper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.texture.TextureManager;
 import net.minecraft.util.ResourceLocation;
+import cpw.mods.fml.client.FMLClientHandler;
 
-import java.util.HashMap;
-
-public abstract class GuiToggleSwitch {
-    class ToggleButton {
+public abstract class GuiToggleSwitch
+{
+    class ToggleButton
+    {
         int u;
         int v;
         String tooltip;
@@ -30,11 +32,13 @@ public abstract class GuiToggleSwitch {
     ResourceLocation texture;
     HashMap<Integer, ToggleButton> buttons = new HashMap<Integer, ToggleButton>();
 
-    public GuiToggleSwitch() {
+    public GuiToggleSwitch()
+    {
         this.mc = FMLClientHandler.instance().getClient();
     }
 
-    public void draw(TextureManager renderEngine) {
+    public void draw(TextureManager renderEngine)
+    {
         renderEngine.bindTexture(texture);
         ToggleButton button = buttons.get(state);
         drawTexturedModalRect(x, y, button.u, button.v, width, height);
@@ -47,52 +51,54 @@ public abstract class GuiToggleSwitch {
             container.drawCreativeTabHoveringText(tooltip, cx + 77 - (tooltipWidth / 2), cy + 100);
     }
 
-    public void setPos(int x, int y) {
+    public void setPos(int x, int y)
+    {
         this.x = x;
         this.y = y;
     }
 
-    public void mouseClicked(int x, int y, int mouseButton) {
-        if (isMoverOver()) {
+    public void mouseClicked(int x, int y, int mouseButton)
+    {
+        if (isMoverOver())
+        {
             onClicked();
         }
     }
 
-    public boolean isMoverOver() {
+    public boolean isMoverOver()
+    {
         mouseX = container.getMouseX();
         mouseY = container.getMouseY();
-        if (mouseX > this.x && mouseX < this.x + width && mouseY > this.y && mouseY < this.y + height) {
+        if (mouseX > this.x && mouseX < this.x + width && mouseY > this.y && mouseY < this.y + height)
+        {
             return true;
         }
         return false;
     }
 
-    public int getState() {
+    public int getState()
+    {
         return this.state;
     }
 
-    private void onClicked() {
+    private void onClicked()
+    {
         this.state++;
         if (this.state == this.numStates)
             this.state = 0;
     }
 
-    /**
-     * Draws a textured rectangle at the stored z-value. Args: x, y, u, v, width, height
-     */
-    public void drawTexturedModalRect(int par1, int par2, int par3, int par4, int par5, int par6) {
+    /** Draws a textured rectangle at the stored z-value. Args: x, y, u, v, width, height */
+    public void drawTexturedModalRect(int par1, int par2, int par3, int par4, int par5, int par6)
+    {
         float var7 = 0.00390625F;
         float var8 = 0.00390625F;
         Tessellator var9 = Tessellator.instance;
         var9.startDrawingQuads();
-        var9.addVertexWithUV((double) (par1 + 0), (double) (par2 + par6), (double) this.zLevel, (double) ((float) (par3 + 0) * var7),
-                (double) ((float) (par4 + par6) * var8));
-        var9.addVertexWithUV((double) (par1 + par5), (double) (par2 + par6), (double) this.zLevel, (double) ((float) (par3 + par5) * var7),
-                (double) ((float) (par4 + par6) * var8));
-        var9.addVertexWithUV((double) (par1 + par5), (double) (par2 + 0), (double) this.zLevel, (double) ((float) (par3 + par5) * var7),
-                (double) ((float) (par4 + 0) * var8));
-        var9.addVertexWithUV((double) (par1 + 0), (double) (par2 + 0), (double) this.zLevel, (double) ((float) (par3 + 0) * var7),
-                (double) ((float) (par4 + 0) * var8));
+        var9.addVertexWithUV(par1 + 0, par2 + par6, this.zLevel, (par3 + 0) * var7, (par4 + par6) * var8);
+        var9.addVertexWithUV(par1 + par5, par2 + par6, this.zLevel, (par3 + par5) * var7, (par4 + par6) * var8);
+        var9.addVertexWithUV(par1 + par5, par2 + 0, this.zLevel, (par3 + par5) * var7, (par4 + 0) * var8);
+        var9.addVertexWithUV(par1 + 0, par2 + 0, this.zLevel, (par3 + 0) * var7, (par4 + 0) * var8);
         var9.draw();
     }
